update blockchains and logos for sunny awards
